Fix ak.from_arrow and unit-scaling (should be floating point) to address last comment.

+++ include/awkward/datetime_util.h
@@ -105,7 +105,7 @@ namespace awkward {
 
     std::tuple<std::string, int64_t> datetime_data(const std::string& format);
 
-    int64_t scale_from_units(const std::string& format, uint64_t index);
+    double scale_from_units(const std::string& format, uint64_t index);
   }
 }
