+++ include/awkward/kernels.h
@@ -2738,7 +2738,7 @@ extern "C" {
     int64_t tooffset,
     const int64_t* fromptr,
     int64_t length,
-    int64_t scale);
+    double scale);
 
   EXPORT_SYMBOL ERROR
   awkward_NumpyArray_getitem_boolean_nonzero_64(
